refactor: exception, config 관련 클래스 패키지 변경 및 이름 변경
