+++ src/main/java/roomescape/auth/application/JwtTokenProvider.java
@@ -5,7 +5,7 @@ import io.jsonwebtoken.io.Decoders;
 import io.jsonwebtoken.security.Keys;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
-import roomescape.exception.IllegalTokenException;
+import roomescape.auth.exception.IllegalTokenException;
 
 import java.security.Key;
 import java.util.Date;
