+++ src/main/java/roomescape/auth/exception/IllegalTokenException.java
@@ -1,0 +1,7 @@
+package roomescape.auth.exception;
+
+public class IllegalTokenException extends AuthorizationException {
+    public IllegalTokenException(String message) {
+        super(message);
+    }
+}
