--- src/main/java/roomescape/exception/IllegalTokenException.java
@@ -1,7 +1,0 @@
-package roomescape.exception;
-
-public class IllegalTokenException extends NotAuthenticatedException {
-    public IllegalTokenException(String message) {
-        super(message);
-    }
-}
