--- src/main/java/roomescape/exception/TokenNotExistException.java
@@ -1,7 +1,0 @@
-package roomescape.exception;
-
-public class TokenNotExistException extends NotAuthenticatedException {
-    public TokenNotExistException() {
-        super("토큰이 존재하지 않습니다.");
-    }
-}
