+++ src/main/java/roomescape/global/config/WebMvcConfiguration.java
@@ -1,4 +1,4 @@
-package roomescape;
+package roomescape.global.config;
 
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
