+++ src/main/java/roomescape/global/dto/ErrorResponse.java
@@ -1,4 +1,4 @@
-package roomescape.exception;
+package roomescape.global.dto;
 
 public record ErrorResponse(String message) {
 }
