+++ src/test/java/roomescape/reservation/application/ReservationTimeServiceTest.java
@@ -7,7 +7,7 @@ import org.mockito.BDDMockito;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
-import roomescape.exception.ViolationException;
+import roomescape.global.exception.ViolationException;
 import roomescape.reservation.domain.ReservationRepository;
 import roomescape.reservation.domain.ReservationTime;
 import roomescape.reservation.domain.ReservationTimeRepository;
