+++ src/test/java/roomescape/reservation/presentation/ReservationControllerTest.java
@@ -12,12 +12,12 @@ import org.springframework.context.annotation.ComponentScan.Filter;
 import org.springframework.context.annotation.FilterType;
 import org.springframework.context.annotation.Import;
 import org.springframework.http.MediaType;
-import roomescape.WebMvcConfiguration;
+import roomescape.global.config.WebMvcConfiguration;
 import roomescape.auth.presentation.AdminAuthorityInterceptor;
 import roomescape.auth.presentation.LoginMemberArgumentResolver;
 import roomescape.common.ControllerTest;
 import roomescape.common.TestWebMvcConfiguration;
-import roomescape.exception.NotFoundException;
+import roomescape.global.exception.NotFoundException;
 import roomescape.reservation.application.ReservationService;
 import roomescape.reservation.application.ReservationTimeService;
 import roomescape.reservation.application.ThemeService;
