+++ src/test/java/roomescape/reservation/presentation/ThemeControllerTest.java
@@ -8,7 +8,7 @@ import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.context.annotation.ComponentScan.Filter;
 import org.springframework.context.annotation.FilterType;
 import org.springframework.http.MediaType;
-import roomescape.WebMvcConfiguration;
+import roomescape.global.config.WebMvcConfiguration;
 import roomescape.auth.presentation.AdminAuthorityInterceptor;
 import roomescape.auth.presentation.LoginMemberArgumentResolver;
 import roomescape.common.ControllerTest;
